+++ app/src/app/examples/dashboards/dashboards.config.js
@@ -10,19 +10,35 @@
         $translatePartialLoaderProvider.addPart('app/examples/dashboards');
 
         $stateProvider
+        .state('triangular.sales-layout', {
+            abstract: true,
+            views: {
+                sidebarLeft: {
+                    templateUrl: 'app/triangular/components/menu/menu.tmpl.html',
+                    controller: 'MenuController',
+                    controllerAs: 'vm'
+                },
+                content: {
+                    template: '<div id="admin-panel-content-view" flex ui-view></div>'
+                },
+                belowContent: {
+                    template: '<div ui-view="belowContent"></div>'
+                }
+            }
+        })
         .state('triangular.admin-default.dashboard-general', {
             url: '/dashboards/general',
-            templateUrl: 'app/examples/dashboards/dashboard-general.tmpl.html'
+            templateUrl: 'app/examples/dashboards/general/dashboard-general.tmpl.html'
         })
         .state('triangular.admin-default.dashboard-analytics', {
             url: '/dashboards/analytics',
-            templateUrl: 'app/examples/dashboards/dashboard-analytics.tmpl.html',
+            templateUrl: 'app/examples/dashboards/analytics/dashboard-analytics.tmpl.html',
             controller: 'DashboardAnalyticsController',
             controllerAs: 'vm'
         })
         .state('triangular.admin-default.dashboard-server', {
             url: '/dashboards/server',
-            templateUrl: 'app/examples/dashboards/dashboard-server.tmpl.html',
+            templateUrl: 'app/examples/dashboards/server/dashboard-server.tmpl.html',
             controller: 'DashboardServerController',
             controllerAs: 'vm'
         })
@@ -32,13 +48,31 @@
         })
         .state('triangular.admin-default.dashboard-social', {
             url: '/dashboards/social',
-            templateUrl: 'app/examples/dashboards/dashboard-social.tmpl.html',
+            templateUrl: 'app/examples/dashboards/social/dashboard-social.tmpl.html',
             controller: 'DashboardSocialController',
             controllerAs: 'vm'
         })
+        .state('triangular.admin-default.dashboard-sales', {
+            url: '/dashboards/sales',
+            data: {
+                layout: {
+                    showToolbar: false
+                }
+            },
+            views: {
+                '': {
+                    templateUrl: 'app/examples/dashboards/sales/dashboard-sales.tmpl.html',
+                    controller: 'DashboardSalesController',
+                    controllerAs: 'vm'
+                },
+                'belowContent': {
+                    templateUrl: 'app/examples/dashboards/sales/fab-button.tmpl.html'
+                }
+            }
+        })
         .state('triangular.admin-default.dashboard-draggable', {
             url: '/dashboards/draggable-widgets',
-            templateUrl: 'app/examples/dashboards/dashboard-draggable.tmpl.html'
+            templateUrl: 'app/examples/dashboards/draggable.tmpl.html'
         });
 
         triMenuProvider.addMenu({
@@ -53,6 +87,10 @@
             },{
                 name: 'MENU.DASHBOARDS.GENERAL',
                 state: 'triangular.admin-default.dashboard-general',
+                type: 'link'
+            },{
+                name: 'MENU.DASHBOARDS.SALES',
+                state: 'triangular.admin-default.dashboard-sales',
                 type: 'link'
             },{
                 name: 'MENU.DASHBOARDS.SERVER',